bq/requirements.txt
google-cloud-bigquery-datatransfer==1 